+++ src/main/groovy/com/cabolabs/openehr/rm_1_0_2/data_types/quantity/DvAmount.groovy
@@ -2,7 +2,7 @@ package com.cabolabs.openehr.rm_1_0_2.data_types.quantity
 
 abstract class DvAmount extends DvQuantified {
 
-   Float accuracy
+   BigDecimal accuracy
    Boolean accuracy_is_percent
 
    // operators
